creates ProcessedImages DB
Use this on your local machine to create an identical instance of the ProcessedImages MySQL database. This should work on *nix systems, if it doesn't I would first check that the script is using the same engine that you should (it's set to use InnoDB right now).
To use: 
in a terminal type 'mysql -u root -p', then once you're logged in type 'source seefood.sql'. 
This assumes that seefood.sql is in your current working directory and has the file permissions to run.